Split native Zenmap models and support code
