+++ zenmap/macos/native/ZenmapSupport.swift
@@ -1,0 +1,40 @@
+import Foundation
+import AppKit
+import Darwin
+
+func openHelpURL(_ string: String) {
+        guard let url = URL(string: string) else {
+            return
+        }
+
+        NSWorkspace.shared.open(url)
+    }
+
+
+
+extension ProcessInfo {
+    var machineHardwareName: String {
+        var systemInfo = utsname()
+        uname(&systemInfo)
+
+        return withUnsafePointer(to: &systemInfo.machine) {
+            $0.withMemoryRebound(to: CChar.self, capacity: 1) {
+                String(cString: $0)
+            }
+        }
+    }
+}
+
+extension JSONEncoder {
+    static var profileEncoder: JSONEncoder {
+        let encoder = JSONEncoder()
+        encoder.outputFormatting = [.prettyPrinted, .sortedKeys]
+        return encoder
+    }
+}
+
+extension JSONDecoder {
+    static var profileDecoder: JSONDecoder {
+        JSONDecoder()
+    }
+}
